+++ gatsby-theme-github-stats/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 - feat: add theme-ui to start theme API for customizing page style
+- fix: Creating tokens in siteMetadata for header & footers.
+- feat: abstracted header and footer to components
 
 # 0.1.0 - July 03, 2019